==> 01_segment_planes.ipynb <==
import os
import numpy as np
import pandas as pd
from skimage.segmentation import clear_border
from skimage.io import imsave
from tifffile import imread
from skimage.morphology import opening,dilation,disk
from skimage.segmentation import find_boundaries

import mxnet
from cellpose import models

# load cellpose model
model = models.Cellpose(gpu=False, model_type='cyto')

path_frames_dir = './data_frames'
path_masks_dir = './segmentation_frames'

## Segmentation of bacteria (Cellpose)

im_list = [x for x in os.listdir(path_frames_dir) if 'ch00' in x]

for im_file in im_list:

    file_path = os.path.join(path_frames_dir,im_file)
    
    # read in the file
    im = imread(file_path)
   
    # segment the selected plane
    labels, flows, styles, diams = model.eval(im, diameter=40, channels=[[0,0]])

    # clear border
    labels_2D = clear_border(labels)

    # morphological operations enhancing the regions
    boundaries = find_boundaries(labels_2D)
    labels_2D[boundaries]=0

    selem_open = disk(6)
    labels_open = opening(labels_2D, selem_open)
    selem_dil = disk(1)
    labels_dilated = dilation(labels_open,selem_dil)

    # save segmentation
    save_path = os.path.join(path_masks_dir,im_file.replace('.tiff','_cells.png'))
    imsave(save_path,labels_dilated.astype('uint16')) 

import napari

viewer = napari.Viewer()

viewer.add_image(im,blending='additive',colormap='gray')
viewer.add_labels(labels_dilated,blending='additive')

## Run ijm for MaxEntropy thresholding
