==> 03_annotate_cells.ipynb <==
import os
import pandas as pd

path_singles_dir = './singles_ch01'
path_manual_selection = './manual_selection'

df_file = './cell_data.pkl' 
df_save = './cell_data_annotated.pkl' 

df = pd.read_pickle(df_file)

ph1_dir = os.path.join(path_manual_selection,'ph_1')
ph2_dir = os.path.join(path_manual_selection,'ph_2')
ph3_dir = os.path.join(path_manual_selection,'ph_3')
mis_dir = os.path.join(path_manual_selection,'missegmented')

ph1_list = os.listdir(ph1_dir)
ph2_list = os.listdir(ph2_dir)
ph3_list = os.listdir(ph3_dir)
mis_list = os.listdir(mis_dir)

df['ann_phase1'] = [f'{x}.tiff' in ph1_list for x in df.cell_name]
df['ann_phase2'] = [f'{x}.tiff' in ph2_list for x in df.cell_name]
df['ann_phase3'] = [f'{x}.tiff' in ph3_list for x in df.cell_name]
df['ann_mis'] = [f'{x}.tiff' in mis_list for x in df.cell_name]

df.loc[df.ann_phase1,'ann'] = 'phase1'
df.loc[df.ann_phase2,'ann'] = 'phase2'
df.loc[df.ann_phase3,'ann'] = 'phase3'
df.loc[df.ann_mis,'ann'] = 'error'

sel_data = df.loc[df.ann==df.ann,:]
# cound annotated objects
len(sel_data)

# drop heavy columns that will not be needed

df = df.drop(['filled_image', 'intensity_image'],axis=1)

df.to_pickle(df_save)
